DEV-766 add gdc app support
This commit adds the initial prototype for gdc apps. The apps
are react-redux wrappers around a react component. Metadata
must be provided by each GDC App. This metadata allows the app
to be tracked and referenced by the gff core.

Example apps have been added to the portal prototype to demonstrate
how the app registration and selection works.