Feat/all apps (#8)
DEV-770 more proto for all apps, views, and styles

Views - This commits adds a loose concept of a "view".
Views can exist for single entities and collections of entities.
Ideally, these views take their input via props and a wrapper
component is used to fetch data from the api.  This should 
allow for more reuse in future.  An example of a wrappers
would be a contextual wrapper that fetches data based on
the current filters.

Styling - The user flow protos are slightly prettier now.  We
are approaching the point where we might need to start
extracting components for a ui kit.  

Colors - A color these has been added to the tailwind configs.
See the colors demo page for more details.