DEV-770 add files to cohort manager
This commit puts the cases view into use and follows the same
pattern for the files view. These are now used in the cohort
manager.

This is the first used of passing api request params to the
data access hooks.  While it works, it's still tightly coupling
the api to the display code. Also, the api shape to the display
data model shape. Both the request and the shape need to be
abstracted to prevent this tight coupling.