DEV-770 many changes for upcoming demo
* DEV-770 add files to cohort manager

This commit puts the cases view into use and follows the same
pattern for the files view. These are now used in the cohort
manager.

This is the first used of passing api request params to the
data access hooks.  While it works, it's still tightly coupling
the api to the display code. Also, the api shape to the display
data model shape. Both the request and the shape need to be
abstracted to prevent this tight coupling.

* DEV-773 DAL ramblings of a mad man

These thoughts are incomplete and need to be organized still.

* DEV-770 refactor projects

This commit moves the project data access to the common gdcapi
code, which also get refactored to be more generic for any gdcapi
endpoint.

Also updated projects view to use the prettier cards from the analysis
views.

* DEV-770 update select

This adds a themed version of react-select.  Once we have a
a gdc theme, we should update that component. Then, we'll want
to use that component instead of react-select directly.

also, replaced the dropdowns in the studies page.

* DEV-770 add gdc themed select

* DEV-770 unify studies

Both flows use the studies view now.

* DEV-770 consolidate apps

Moved apps into one place and shared between all-apps and many-pages.
App cards now have front and backs.  This is rather rudimentary.
We'll want to refactor this into something nicer later.

* DEV-770 hook up repository view

* added mouseover text to apps (#10)

* added mouseover text to apps. cohort viewer pending

* added mouseover text for cohort viewer

* adjusted text on buttons and cohort bar; make oncogrid load for all the apps in many pages (#11)

* adjusted buttons text in studies page; made oncogrid image load for every app in many pages

* edited text and order of cohorts in cohort bar dropdown

* removed stray semicola

* formatting

Co-authored-by: Phil Tom <philtom@uchicago.edu>

* DEV-770 open oncogrid and scroll

* DEV-770 checkbox on studies

Co-authored-by: wteouchicago <73256434+wteouchicago@users.noreply.github.com>