allow prototype to run in production mode
A variety of linting errors needed to be resolved.  The
tailwind configs needed to be fixed for tree-shaking to work
correctly. This allows `npm run build` to successfully complete.
The production version of the prototype is much faster than the
dev version.

Also made some layout and display tweaks to the prototype.