PEAR-535 add base path
The next.js configuration for basePath has been set to /v2 to
support the deployment of portal v2 in conjunction with portal v1.

Since the image component from next/image does not support basePath,
a wrapper component has been added to automatically add the basePath
when appropriate. This removes the need to manually add the base
path to image sources.