PEAR-862 move to relative urls for prod
While setting the uat environment at uat-portal.gdc.cancer.gov,
we realized that the prod portal and api urls were still being
used. This is because we use absolute urls for backend
communications.

To overcome this, I updated the urls to be relative and for all
api calls to pass through the portal api proxy at /auth/api/.
This will allow the prod portal to run on any domain.

Using the same domain will also avoid CORS issues, such as the
ones we saw with cohort api.

The absolution urls for the portal v1 annotations pages were
left as-is. These will eventually be replaced by a v2 annotations
implementation.