PEAR-1020 reconfigure datadog rum (#386)
Since data rum is intended monitoring browser activity, it belongs in
the portal package, not the core package.

This commit also updates dd rum to the latest version with the latest
configs.  Additional features are enabled.  We can evaluate if they are
worth keeping during the uat phase of this project.  Session recording
was disabled to avoid issues with controlled data.